Add additional checking on Customer Portal creating of views if the ECommerce is active.

+++ Integration/dbo/Views/vwCPBillingAccountPayments.sql
@@ -1,4 +1,5 @@
-﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vwCPBillingAccountPayments')
+﻿/*
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vwCPBillingAccountPayments')
 	DROP VIEW vwCPBillingAccountPayments
 GO
 
@@ -13,4 +14,25 @@ SELECT
   FROM vwCPPaymentsDetails
   GROUP BY dblAmount, dtmDate, strCheckNo,strCustomerNo
   --order by dtmDate
+GO
+*/
+
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vwCPBillingAccountPayments')
+	DROP VIEW vwCPBillingAccountPayments
+GO
+
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'vwCPPaymentsDetails')
+	EXEC('
+		CREATE VIEW [dbo].[vwCPBillingAccountPayments]
+		AS
+		SELECT
+			   id = row_number() over (order by dblAmount)
+			  ,strCustomerNo
+			  ,dtmDate
+			  ,strCheckNo
+			  ,dblAmount = sum(dblAmount)
+		  FROM vwCPPaymentsDetails
+		  GROUP BY dblAmount, dtmDate, strCheckNo,strCustomerNo
+		  --order by dtmDate
+	')
 GO
